Mxonline362/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.12
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.6
jdcal==1.4
kombu==4.1.0
MySQL-python==1.2.5
mysqlclient==1.3.7
odfpy==1.3.6
openpyxl==2.5.2
pbr==4.0.2
Pillow==5.1.0
PyMySQL==0.8.0
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.10.0
stevedore==1.28.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
